requirements.txt
psycopg2-binary==2.9.9
reportlab==4.0.7